Separate Query run logic from manager
